requirements.txt
python-telegram-bot==20.4
anthropic==0.8.0
google-auth==2.26.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
pytest==7.4.3